requirements.txt
google-generativeai
duckduckgo-search
python-dotenv
colorama
